added missing animation
